+++ packages/teamscale-javascript-instrumenter/test/system/Instrumenter.test.ts
@@ -37,6 +37,43 @@ test('Instrumentation: React App. Vendor.js', () => {
 	).resolves.toHaveProperty('translated', 1);
 });
 
+test('Instrumentation: React App. Vendor.js. Skip its instrumentation based on pattern.', () => {
+	const outputDir = path.join(__dirname, '..', '..', 'outputs');
+	return expect(
+		App.runForConfigArguments({
+			inputs: [path.join(__dirname, 'inputs', 'vite-react-vendor.js')],
+			in_place: false,
+			exclude_bundle: ['**/*-vendor.js'],
+			collector: 'localhost:54321',
+			to: outputDir
+		})
+	).resolves.toHaveProperty('excluded', 1);
+});
+
+test('Instrumentation: React App. Vendor.js; single quoted collector config; TS-33372', () => {
+	const outputDir = path.join(__dirname, '..', '..', 'outputs');
+	return expect(
+		App.runForConfigArguments({
+			inputs: [path.join(__dirname, 'inputs', 'vite-react-vendor.js')],
+			in_place: false,
+			collector: '\'localhost:54321\'',
+			to: outputDir
+		})
+	).resolves.toHaveProperty('translated', 1);
+});
+
+test('Instrumentation: React App. Vendor.js; double quoted collector config; TS-33372', () => {
+	const outputDir = path.join(__dirname, '..', '..', 'outputs');
+	return expect(
+		App.runForConfigArguments({
+			inputs: [path.join(__dirname, 'inputs', 'vite-react-vendor.js')],
+			in_place: false,
+			collector: '"localhost:54321"',
+			to: outputDir
+		})
+	).resolves.toHaveProperty('translated', 1);
+});
+
 test('Instrumentation: Angular App', () => {
 	const outputDir = path.join(__dirname, '..', '..', 'outputs');
 	return expect(